Write stub and columns label

Write stub and columns label

write_stub_head_and_column_labels(
  gt_table,
  ordered_gt_data,
  wb,
  sheet_name,
  row_to_start
)

Arguments

gt_table

A gt table

ordered_gt_data

ordered data from `create_ordered_data`

wb

Name of Workbook object like `openxlsx::createWorkbook()`

sheet_name

name of the sheet in the workbook

row_to_start

Number of the row where to start writting the stub and body.

Value

Nothing. It will update the wb object with stub and columns labels.

Examples

wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, "gtcars_example")

ordered_example <- gtopenxlsx::gtcars_8 |>
  gt::gt() |>
  create_ordered_data()
gt_table_example <- gtopenxlsx::gtcars_8 |>
  gt::gt()
ordered_example <- gt_table_example |>
  create_ordered_data()
gt_table_example |>
  write_stub_head_and_column_labels(ordered_gt_data = ordered_example, wb, "gtcars_example", 1)